requirement.txt
Django==4.2.1
djangorestframework==3.14.0
geopy==2.4.1
